Refactoring: move common CLI functions to doozerlib/cli/__init__.py
This will allow to split individual commands into separate files.
